move drag setup from component to drag service
closes #238
