+++ projects/ngx-d3timeline/src/lib/content/activity-rectangle.component.ts
@@ -7,10 +7,6 @@ import {
   ChangeDetectionStrategy
 } from '@angular/core';
 import { ActivityRectangle } from './activity-rectangle';
-import { drag } from 'd3-drag';
-import { select, event } from 'd3-selection';
-import { Store } from '../store/store';
-import * as fromActions from '../store/actions';
 import { ActivityDragService } from './activity-drag.service';
 
 @Component({
@@ -36,10 +32,7 @@ export class ActivityRectangleComponent implements AfterViewInit {
 
   @ViewChild('activityRectangleEl') activityRectangleEl: ElementRef;
 
-  constructor(
-    private store: Store,
-    private activityDragService: ActivityDragService
-  ) {}
+  constructor(private activityDragService: ActivityDragService) {}
 
   ngAfterViewInit() {
     this.setupDrag();
@@ -47,34 +40,10 @@ export class ActivityRectangleComponent implements AfterViewInit {
 
   private setupDrag() {
     if (this.activityRectangleEl) {
-      const onDrag = drag()
-        .on('start', this.onDragStarted.bind(this))
-        .on('drag', this.onDragging.bind(this))
-        .on('end', this.onDragEnded.bind(this));
-
-      onDrag(select(this.activityRectangleEl.nativeElement));
+      this.activityDragService.setupDrag(
+        this.activityRectangle,
+        this.activityRectangleEl.nativeElement
+      );
     }
-  }
-
-  private onDragStarted() {
-    this.store.dispatch(
-      new fromActions.TimelineDragStartedAction({
-        activityRectangle: this.activityRectangle,
-        event
-      })
-    );
-  }
-
-  private onDragging() {
-    this.store.dispatch(
-      new fromActions.TimelineDraggingAction({
-        activityRectangle: this.activityRectangle,
-        event
-      })
-    );
-  }
-
-  private onDragEnded() {
-    this.activityDragService.onDragEnd();
   }
 }
